EUI-3472 squashed commit
EUI-3472 hardcoded placeholder


EUI-3472 pseudo guard and refinement of components


EUI-3472 feature toggling and trying to fix tests
