+++ src/components/banners/alert/alert.component.ts
@@ -3,7 +3,8 @@ import { Component, Input } from '@angular/core';
 enum AlertMessageType {
   WARNING = 'warning',
   SUCCESS = 'success',
-  ERROR = 'error'
+  ERROR = 'error',
+  INFORMATION = 'information'
 }
 
 @Component({
@@ -19,6 +20,7 @@ export class AlertComponent {
   public static readonly TYPE_WARNING = 'warning';
   public static readonly TYPE_SUCCESS = 'success';
   public static readonly TYPE_ERROR = 'error';
+  public static readonly TYPE_INFORMATION = 'information';
 
   @Input()
   public type: AlertMessageType;
