+++ src/shared/components/case-viewer/case-challenged-access-view/index.ts
@@ -1,0 +1,1 @@
+export * from './case-challenged-access-view.component';
