+++ src/shared/components/case-viewer/case-full-access-view/case-full-access-view.component.scss
