+++ src/shared/components/case-viewer/index.ts
@@ -1,6 +1,7 @@
 export * from './services';
 export * from './case-event-trigger';
 export * from './case-view';
+export * from './case-challenged-access-view';
 export * from './printer';
 export * from './case-viewer.component';
 export * from './case-viewer.module';
